score_writer/requirements.txt
confluent-kafka==2.10.0
psycopg2-binary==2.9.9
